more automatic tests

+++ automatic/automatic.py
@@ -1,7 +1,8 @@
 #!/usr/bin/env python3
 import sys
-import googletrans
 from googletrans import Translator
+import googletrans
+import os.path
 
 # from collections import namedtuple
 # Result = namedtuple('result', ["src", "dest", "text"])
@@ -32,45 +33,47 @@ translator = Translator()
 
 languages = googletrans.LANGUAGES
 for languageCode, languageName in languages.items():
-    print(languageName)
-    result = translator.translate(englishSource, src='en', dest=languageCode)
-
     outputFileName = "Test_Automatic_"+languageName.capitalize()+".swift"
-    with open(outputFileName, 'w') as outputFile:
-        outputFile.write("//\n")
-        outputFile.write("//  "+outputFileName+"\n")
-        outputFile.write("//\n")
-        outputFile.write("//  Created by automatic.py\n")
-        outputFile.write("//\n")
-        outputFile.write("\n")
-        outputFile.write("import XCTest\n")
-        outputFile.write("\n")
-        outputFile.write("final class TranslationsTestsAutomatic"+languageName.capitalize()+": XCTestCase {\n")
-        outputFile.write("\n")
-        outputFile.write("    func test_automatic_"+languageName+"() {\n")
-        outputFile.write("        let "+languageName+" = Translate"+languageName.capitalize()+"()\n")
-        outputFile.write("        "+languageName+".groupSeparator = \"\"\n")
-        outputFile.write("        "+languageName+".decimalSeparator = \".\"\n")
+    if os.path.exists(outputFileName):
+        print(languageName+" already exists")
+    else :
+        print(languageName)
+        result = translator.translate(englishSource, src='en', dest=languageCode)
+        with open(outputFileName, 'w') as outputFile:
+            outputFile.write("//\n")
+            outputFile.write("//  "+outputFileName+"\n")
+            outputFile.write("//\n")
+            outputFile.write("//  Created by automatic.py\n")
+            outputFile.write("//\n")
+            outputFile.write("\n")
+            outputFile.write("import XCTest\n")
+            outputFile.write("\n")
+            outputFile.write("final class TranslationsTestsAutomatic"+languageName.capitalize()+": XCTestCase {\n")
+            outputFile.write("\n")
+            outputFile.write("    func test_automatic_"+languageName+"() {\n")
+            outputFile.write("        let "+languageName+" = Translate"+languageName.capitalize()+"()\n")
+            outputFile.write("        "+languageName+".groupSeparator = \"\"\n")
+            outputFile.write("        "+languageName+".decimalSeparator = \".\"\n")
 
-        if languageName == "german":
-            outputFile.write("        "+languageName+".useSoftHyphen = false\n")
-            outputFile.write("        "+languageName+".capitalisation = false\n")
+            if languageName == "german":
+                outputFile.write("        "+languageName+".useSoftHyphen = false\n")
+                outputFile.write("        "+languageName+".capitalisation = false\n")
 
-        for line in result.text.split("\n"):
-            splitLine = line.split(" ", 1)
-            if not line == "":
-                if not len(splitLine) == 2:
-                    print("Error: line not two components:")
-                    print("XXX"+line+"XXX")
-                    exit(0)
-                number = splitLine[0]
-                number = number.replace(" ", "")
-                number = number.replace(",", "")
-                translation = splitLine[1]
-                translation = translation.replace(" ", "")
-                translation = translation.replace(",", "")
-                outputFile.write("        XCTAssertEqual("+languageName+".translate("+number+"), \""+translation+"\")\n")
-        outputFile.write("        }\n")
-        outputFile.write("}\n")
+            for line in result.text.split("\n"):
+                splitLine = line.split(" ", 1)
+                if not line == "":
+                    if not len(splitLine) == 2:
+                        print("Error: line not two components:")
+                        print("XXX"+line+"XXX")
+                    else:
+                        number = splitLine[0]
+                        number = number.replace(" ", "")
+                        number = number.replace(",", "")
+                        translation = splitLine[1]
+                        translation = translation.replace(" ", "")
+                        translation = translation.replace(",", "")
+                        outputFile.write("        XCTAssertEqual("+languageName+".translate("+number+"), \""+translation+"\")\n")
+            outputFile.write("        }\n")
+            outputFile.write("}\n")
 
-    outputFile.close()
+        outputFile.close()
